backend/requirements.txt
# backend/requirements.txt
django
psycopg2-binary
google-generativeai
django-cors-headers
djangorestframework
django-filter
djangorestframework-simplejwt